libsysprof-profiler: buffer controlfd data to intermediate memfd
This ensures that we don't clobber things by writing them without possible
translation. That is problematic when each controlfd process might have
it's own counter registration or JIT maps as they could collide with the
main writer's own state.

This fixes some situations where the counter data was coming in with
erroneous values.