libsysprof-gtk: try harder to get motion events
I still see some occasionaly hiccups in the GtkEventControllerMotion where
if you move slowly from the SysprofTimeRuler area to the headerbar, you
might not get a GtkEventControllerMotion::leave event triggered. But we
can deal with that later as that's clearly something we should be able to
expect to be triggered.