sysprof: add SysprofTimeFilterModel
This is a model that expects the set to be in sorted order by time, and
requires that the items are a SysprofDocumentFrame. If that holds true,
it can binary search to find the correct frames based on a requested
time-span that is inclusive only of matching frames.

This is useful to dive down into data based on either the visible or
selected time span of a session without having to expensive GtkFilter
operations (which would look at every object in the set).