sysprof: filter the callgraph by the selected time span
This ensures the callgraph code has fewer items to look at. The first level
of filtering is done by the SysprofDocument bitset index for items
matching the type (via SysprofDocumentBitsetIndex). Then the new
SysprofTimeFilterModel filters the frames by time-span by binary searching
for the first and last indexes which match that time span, avoiding the
pathological case of linear search for the spans (which would inflate a
GObject for every element).