sysprof: start on symbol categorization
This will end up getting moved a bit to CallgraphNode so that we can
look at the parents, but for now just do this so I can test things out
a bit and see how we like it.

Eventually, the goal would be to categorize the callgraph and show that
information in the sidebar for quick overviews of where time is spent.

It's a bit laborious to keep up and add symbols for this sort of stuff,
but I think the "at a glance" value is pretty huge.