sysprof: fixed row height and text labels
This makes things look a bit more like flamegraphs.pl in the sense that we
have some labels and separation between rows. Also, use a ScrolledWindow so
that we can have much taller graphs to accommodate deep stack traces.

We might want to jump to the bottom at some point, but this gets things in
place for now. Icicle graphs are another option (invert).