sysprof: scroll to the bottom of the flame graph
Unfortunately we don't have a good way to do this in GtkScrolledWindow
where you have an initial gravity pulling you to the bottom of a viewport.

We might be able to fake it with a focusable widget, but that sounds like
more work than just animating to the bottom of the viewport.

Related #93