manifest: ensure we have access to ~/.local/share/flatpak/
That is the common path for a --user installation, and it gets filtered
from us even if we have --filesystem=home. Specifically enforce adding it
in our manifest so that we can read it.

This seems to fix loading a bunch of symbols from Flatpak'd Sysprof for me.

Fixes GNOME/gnome-builder#2090