sysprof/greeter: load/save recording template for persistence
This sets things up to save the state to disk if the greeter suceeded at
creating a profiler instance so that it may be used again on the next
run. Additionally, when loading the greeter it attempts to load from disk
the saved state.

We may want to extend this in the future to allow using sysprof-cli with
a recording template instead of command line options.