libsysprof-capture: Cut off first 8 bytes of initial counter id
This ensures that the id won't be screwed by the 32 → 24 bit conversion
that happens implicitly when using SysprofCaptureCounter & family of
structs.

These structs are tightly packed and declare the id field as a 24-bit
unsigned, which doesn't necessarily match the counting that is done
via SysprofCollector.

This is by no means a definitive solution, but this gets things in a
more usable state until a better solution is implemented.