Make elf loader search for debug links in .debug dirs
When the distribution is built using yocto project, it's possible
to specify 'dbg-pkgs' image feature, see:
https://docs.yoctoproject.org/dev/ref-manual/features.html
This image feature makes files with package debug symbols being
placed under .debug directories accompanying the original binary
directories and mentioned in the .gnu_debuglink of the original binary.
Effectively the symbols of e.g. /x/y/libz.so will be placed in
/x/y/.debug/libz.so.

This change makes sysprof's elf loader to search for debug links
in such a .debug directories.