sysprof: Sample marks after filtering by time
In the Mark Chart view, each row has a small graph with a filtered and
sampled preview of the marks. The sampling is applied after filtering.
This causes the rows to look increasingly sparse the more you zoom in
the timeline.

Swap the order of the operations. First filter by time range, then
sample the resulting dataset. This turns out to not be a performance
hit, because filtering is merely a couple of O(log N) operations, and
sampling is just an integer game.